// apps/api/src/app/rest-endpoint.model.ts
export enum RestEndpoint {
  TopStories = 'top-stories',
  SubscriberOnly = 'subscriber-only',
  World = 'world',
  Politics = 'politics',
  Ireland = 'ireland',
  Business = 'business',
  MostRead = 'most-read',
  Coronavirus = 'coronavirus',
  Galleries = 'galleries',
  Latest = 'latest',
  Sport = 'sport',
  Culture = 'culture',
  Opinion = 'opinion',
  Brexit = 'brexit',
  Abroad = 'abroad',
  Video = 'video',
  LifeStyle = 'life-style',
}